Changed file name to reflect the class inside it
